Lesson1

+++ Lesson1/triangle_square.rb
@@ -1,7 +1,7 @@
 p "Основание треугольника:"
-foundation = gets.chomp.to_i
+foundation = gets.chomp.to_f
 
 p "Высота треугольника:"
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
 p "Площадь треугольника - #{(foundation * height)/2}"
